Fixed DMA time comparison.

+++ frc971/wpilib/dma_edge_counting.h
@@ -165,7 +165,7 @@ class DMASynchronizer {
   ::std::vector<DMASampleHandlerInterface *> handlers_;
 
   // The time at which we most recently read the sensor values.
-  double sample_time_ = 0.0;
+  uint32_t sample_time_ = 0.0;
 
   DISALLOW_COPY_AND_ASSIGN(DMASynchronizer);
 };
